requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-dotenv==1.2.2
pydantic==2.12.5
requests==2.32.3
twelvelabs==1.2.1
python-multipart>=0.0.9
